+++ src/lib.rs
@@ -9,6 +9,7 @@
 //!   to the well-defined directory.
 //! - Using methods under [`mod@flag`] to use thread-safe, lock-free flags.
 //! - Using methods under [`mod@slack`] to send notifications to Slack with Slack Hook.
+//! - Using methods under [`mod@message`] to subscribe and publish based on Nanomsg.
 //!
 //! These utilities are especial for [wmjtyd/crypto-market](https://github.com/wmjtyd/crypto-market);
 //! however, you can also use it in your project. We licensed it under `Apache-2.0`, the same
@@ -24,3 +25,6 @@ pub mod flag;
 
 #[cfg(feature = "slack")]
 pub mod slack;
+
+#[cfg(feature = "message")]
+pub mod message;